#1131 Rename kt-extensions to kotlin-extensions
