+++ hartshorn-assembly/pom.assembly.xml
@@ -71,7 +71,7 @@
         </dependency>
         <dependency>
             <groupId>org.dockbox.hartshorn</groupId>
-            <artifactId>hartshorn-launchpad-kt-extensions</artifactId>
+            <artifactId>hartshorn-launchpad-kotlin-extensions</artifactId>
         </dependency>
         <dependency>
             <groupId>org.dockbox.hartshorn</groupId>
