+++ hartshorn-launchpad-kotlin-extensions/src/main/kotlin/org/dockbox/hartshorn/inject/provider/ComponentProviderExtensions.kt
@@ -1,0 +1,5 @@
+package org.dockbox.hartshorn.inject.provider
+
+import kotlin.reflect.KClass
+
+inline fun <reified T : Any> ComponentProvider.get(): T = this.get(T::class.java)
