+++ hartshorn-launchpad-samples/basic-hello-world-kotlin/src/main/kotlin/org/dockbox/sample/kotlin/KotlinLaunchpadDemo.kt
@@ -1,6 +1,5 @@
 package org.dockbox.sample.kotlin
 
-import org.dockbox.hartshorn.inject.provider.get
 import org.dockbox.hartshorn.launchpad.HartshornApplication
 
 fun main(args: Array<String>) {
